Split workspace runtime status helpers
